#78. universe_render_foundation: Rename include directory.
