+++ executable/application_framework_examples/unit/deferred_render/example.c
@@ -12,13 +12,12 @@
 #include <kan/inline_math/inline_math.h>
 #include <kan/log/logging.h>
 #include <kan/precise_time/precise_time.h>
-#include <kan/render_foundation/program.h>
-#include <kan/render_foundation/render_graph.h>
-#include <kan/render_foundation/texture.h>
 #include <kan/resource_pipeline/meta.h>
-#include <kan/stream/random_access_stream_buffer.h>
 #include <kan/test_expectation/test_expectation.h>
 #include <kan/universe/macro.h>
+#include <kan/universe_render_foundation/program.h>
+#include <kan/universe_render_foundation/render_graph.h>
+#include <kan/universe_render_foundation/texture.h>
 #include <kan/universe_resource_provider/universe_resource_provider.h>
 #include <kan/universe_time/universe_time.h>
