+++ unit/universe_render_foundation/kan/universe_render_foundation/texture.c
@@ -1,9 +1,9 @@
 #include <kan/cpu_profiler/markup.h>
 #include <kan/log/logging.h>
-#include <kan/render_foundation/render_graph.h>
 #include <kan/render_foundation/resource_texture.h>
-#include <kan/render_foundation/texture.h>
 #include <kan/universe/macro.h>
+#include <kan/universe_render_foundation/render_graph.h>
+#include <kan/universe_render_foundation/texture.h>
 #include <kan/universe_resource_provider/universe_resource_provider.h>
 
 KAN_LOG_DEFINE_CATEGORY (render_foundation_texture);
